refactor: piece 데이터 초기화하는 로직 ChessService로 옮김
